move web app setting panel to runner package
